4. Even Times
Create a program that prints a number from a collection, which appears an even number of times in it. On the first line, you will be given n – the count of integers you will receive. On the next n lines, you will be receiving the numbers. It is guaranteed that only one of them appears an even number of times. Your task is to find that number and print it in the end. 
Examples
Input	Output
3
2
-1
2	2
5
1
2
3
1
5	1