Add files via upload
2. Line Numbers
Write a program that reads a text file (e. g. text.txt) and inserts line numbers in front of each of its lines and count all the letters and punctuation marks. The result should be written to another text file (e. g. output.txt). Use the static class File to read and write all the lines of the input and output files.
Note: use the following structure:
public class LineNumbers
{
    public static void ProcessLines(string inputFilePath, string outputFilePath)
    {
    }
}
Examples
text.txt	output.txt
-I was quick to judge him, but it wasn't his fault.
-Is this some kind of joke?! Is it?
-Quick, hide here. It is safer.	Line 1: -I was quick to judge him, but it wasn't his fault. (37)(4)
Line 2: -Is this some kind of joke?! Is it? (24)(4)
Line 3: -Quick, hide here. It is safer. (22)(4)