Add files via upload
5. Copy Directory
Write a method, which copies a directory with files (without its subdirectories) to another directory. The input folder and the output folder should be given as parameters from the console. If the output folder already exists, first delete it (together with all its content).
Note: use the following structure:
namespace CopyDirectory
{
    using System;

    public class CopyDirectory
    {
        static void Main()
        {
            string inputPath = @$"{Console.ReadLine()}";
            string outputPath = @$"{Console.ReadLine()}";

            CopyAllFiles(inputPath, outputPath);
        }

        public static void CopyAllFiles(string inputPath, string outputPath)
        {
        }
    }
}