Create 3. Songs
Define a class called Song that will hold the following information about some songs:
•	Type List
•	Name
•	Time
Input / Constraints
•	On the first line, you will receive the number of songs - N.
•	On the next N lines, you will be receiving data in the following format:  "{typeList}_{name}_{time}".
•	On the last line, you will receive Type List or "all".
Output
If you receive Type List as an input on the last line, print out only the names of the songs, which are from that Type List. If you receive the "all" command, print out the names of all the songs.