Create 6. Summation Pairs
6. Summation Pairs
The task is not included in the Judge system.
You will receive a sequence of numbers (unique integers) separated by space on the first line. On the second line, you'll receive a target number. Your task is to find the pairs of numbers whose sum equals the target number. For each found pair print "{number} + {number} = {target_number}". You should NOT use the same element twice to fulfill the condition above.
Can you come up with an algorithm that has less time complexity?
Examples
Input	Output
1 5 4 2 3 0
4	1 + 3 = 4
4 + 0 = 4
11 8 5 6 9 2 7 3 4
11	5 + 6 = 11
9 + 2 = 11
8 + 3 = 11
7 + 4 = 11
Hints
First, we read the sequence of numbers and the target number from the console:
 
Then, we write nested for-loops to loop through the list of numbers and check the sum of each two numbers with the target:
  
That is not enough. When we find a matching pair of numbers, we should find a way to eliminate them from the following summation. One way to do that is by changing the value of the element and continuing the loop when we hit that value:
 
This is an example solution to our problem. 
However, is it possible to improve the solution, so the result is found faster? We can use just one loop to iterate over the sequence and an additional set, where we will keep the difference between the target and each of the numbers in the list. Then we will continue to look for exactly that number from the following numbers in the given sequence.
First, let us create the set to keep the numbers as described above. Then, create an empty dictionary that will keep the same number (as added in the set) as a key, and the list's number (that we subtracted from the target) as a value:
 
Next, iterate over the sequence of numbers and start subtracting each of them from the target number. The resulting number should be added to the set, and the value should be added in the dictionary mapped to the resulting number (as key):
 
To create a match where the sum of two numbers should be equal to the target number, we should check if any of the next numbers are in the target set. If the condition is met, print the value and its pair (the value from the dictionary mapping the same key). Do not forget to remove the found pair from the set and the dictionary:
 
This is a faster way to solve the problem. 😊