Create 4. Symbol in Matrix
4. Symbol in Matrix
Write a program that reads N, a number representing the rows and cols of a matrix. On the next N lines, you will receive rows of the matrix. Each row consists of ASCII characters. After that, you will receive a symbol. Find the first occurrence, starting from the top left, of that symbol in the matrix and print its position in the format: "({row}, {col})". If there is no such symbol print an error message "{symbol} does not occur in the matrix"
Examples
Input	Output
3
ABC
DEF
X!@
!	(2, 1)
4
asdd
xczc
qwee
qefw
4	4 does not occur in the matrix